% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Regression.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Fit a Linear Regression Model}
\usage{
fit_linear_regression(Y, X)
}
\arguments{
\item{Y}{: A numeric column obtained from a dataframe representing the response variable.}

\item{X}{: A few columns obtained from a dataframe representing predictors.}
}
\value{
A list containing the regression coefficients and other model statistics.
}
\description{
This function fits a linear regression model to the given response \code{Y} and predictors \code{X}.
}
\examples{
# Example 1: Simple linear regression
Y = c(2, 4, 6, 8)
X = c(0.5, 0.7, 1.6, 2.2)
fit_linear_regression(Y, X)

# Example 2: Cars93 data
testdata = MASS::Cars93
fit_linear_regression(testdata$Price,testdata[,c("Horsepower","Luggage.room")])

}
