% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Regression.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Fit a Linear Regression Model}
\usage{
fit_linear_regression(Y, X)
}
\arguments{
\item{Y}{: A numeric column obtained from a dataframe representing the response variable.}

\item{X}{: A few columns obtained from a dataframe representing predictors.}
}
\value{
A list containing the regression coefficients and other model statistics.
}
\description{
This function fits a linear regression model to the given response \code{Y} and predictors \code{X}.
}
\examples{
Y <- c(1, 2, 3)
X <- matrix(c(1, 2, 3), ncol = 1)
fit <- fit_linear_regression(Y, X)
print(fit)
}
